feat：技能的查询、上传、下载目录统一优化： 1.针对数字员工（个人助理or数字员工），技能的查询、上传、下载路径都是：byclaw-{userCode}/by/.openclaw/workspace-baiying-agent-{resourceId}/skills/ 2.针对超级助手，技能的查询、上传、下载路径，都是：byclaw-{userCode}/by/.openclaw/workspace/skills/
