+++ byclaw-be/src/main/java/com/iwhalecloud/byai/manager/interfaces/controller/digitemploy/DigitalEmployeeController.java
@@ -122,6 +122,7 @@ public class DigitalEmployeeController {
         EmployeeIdDTO employeeIdDTO = new EmployeeIdDTO();
         employeeIdDTO.setResourceId(ssResource.getResourceId());
         DigitalEmployeeDetailsDTO details = digitalEmployeeApplicationService.findDetailsById(employeeIdDTO);
+        digitalEmployeeApplicationService.applyInputRuntimeFieldsForResponse(details, digitalEmployeeDTO);
 
         return ResponseUtil.successResponse(I18nUtil.get("digemployee.update.success"), details);
     }
