+++ byclaw-be/src/main/java/com/iwhalecloud/byai/state/domain/resource/service/ToolManService.java
@@ -56,8 +56,6 @@ import com.iwhalecloud.byai.manager.entity.auth.PrivilegeGrant;
 import com.iwhalecloud.byai.state.application.service.dataset.DatasetApplicationService;
 import com.iwhalecloud.byai.common.login.auth.CurrentUserHolder;
 import com.iwhalecloud.byai.common.util.CurlParser;
-import com.iwhalecloud.byai.common.util.CurlParser.ParsedCurl;
-import com.iwhalecloud.byai.manager.domain.aimodel.service.AIService;
 import com.iwhalecloud.byai.state.domain.resource.dto.CurlParseResult;
 import com.iwhalecloud.byai.state.domain.resource.dto.ObjectZipImportItem;
 import com.iwhalecloud.byai.state.domain.resource.dto.ObjectZipImportResult;
@@ -71,12 +69,6 @@ import com.iwhalecloud.byai.state.domain.resource.dto.ResourceCurlRunRequest;
 import com.iwhalecloud.byai.state.domain.resource.dto.ResourceCurlRunResult;
 import com.iwhalecloud.byai.state.domain.resource.dto.ToolSaveRequest;
 import com.iwhalecloud.byai.state.domain.sys.service.SequenceService;
-import okhttp3.MediaType;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.RequestBody;
-import okhttp3.Response;
-import okhttp3.ResponseBody;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
@@ -89,19 +81,15 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.util.*;
 import java.util.function.Supplier;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
-import java.util.concurrent.TimeUnit;
 
 @Service
 public class ToolManService {
@@ -191,7 +179,7 @@ public class ToolManService {
     private ResourceTargetJsonBuilder resourceTargetJsonBuilder;
 
     @Autowired
-    private AIService aiService;
+    private ResourceCurlService resourceCurlService;
 
     /** 新版工具资源 JSON 导入允许的 resourceBizType */
     private static final Set<String> IMPORT_TOOL_JSON_NEW_BIZ_TYPES = Set.of(
@@ -220,40 +208,15 @@ public class ToolManService {
     private static final int MAX_IMPORT_BUNDLE_RESOURCE_COUNT = 20;
     private static final String OBJECT_IMPORT_RESOURCE_SUBDIR = "object";
     private static final String VIEW_IMPORT_RESOURCE_SUBDIR = "view";
-    private static final String CURL_GENERATE_SOURCE_RULE = "RULE";
-    private static final String CURL_GENERATE_SOURCE_LLM = "LLM";
-    private static final Pattern URL_PATTERN = Pattern.compile("https?://[^\\s\"'<>]+", Pattern.CASE_INSENSITIVE);
-    private static final Set<String> TOOL_CURL_BIZ_TYPES = Set.of(
-            ResourceBizType.TOOL.getCode(),
-            ResourceBizType.TOOLKIT.getCode(),
-            ResourceBizType.MCP.getCode(),
-            ResourceBizType.AGENT.getCode());
-    private static final OkHttpClient RESOURCE_CURL_HTTP_CLIENT = new OkHttpClient.Builder()
-            .connectTimeout(10, TimeUnit.SECONDS)
-            .readTimeout(30, TimeUnit.SECONDS)
-            .writeTimeout(10, TimeUnit.SECONDS)
-            .build();
+
+
+    // ==================== Curl 调试 ====================
 
     /**
      * 阶段一：解析 curl，返回结构化预览结果（不入库）
      */
     public CurlParseResult parseCurl(String curlCommand) {
-        ParsedCurl parsed = CurlParser.parse(curlCommand);
-
-        CurlParseResult result = new CurlParseResult();
-        result.setResourceName(CurlParser.generateToolName(parsed.getPath()));
-        result.setResourceDesc("");
-        result.setMethod(parsed.getMethod().toLowerCase());
-        result.setUrl(parsed.getBaseUrl() + parsed.getPath());
-        result.setUrlOri(parsed.getFullUrl());
-        result.setCurlRaw(curlCommand);
-        result.setBodyParams(CurlParser.extractBodyParams(parsed.getBody()));
-        result.setQueryParams(CurlParser.extractQueryParams(parsed.getQueryParams()));
-        result.setPathParams(new ArrayList<>());
-        result.setHeaderParams(CurlParser.extractHeaderParams(parsed.getHeaders()));
-
-
-        return result;
+        return resourceCurlService.parseCurl(curlCommand);
     }
 
     /**
@@ -263,27 +226,7 @@ public class ToolManService {
      * @date 2026-05-08 00:00:00
      */
     public ResourceCurlGenerateResult generateResourceCurl(ResourceCurlGenerateRequest request) {
-        Long resourceId = request == null ? null : request.getResourceId();
-        ResourceCurlContent content = loadResourceCurlContent(resourceId);
-        if (StringUtils.isBlank(content.getSourceContent())) {
-            throw new IllegalArgumentException("资源sourceContent为空，无法生成curl脚本");
-        }
-
-        ResourceCurlGenerateResult result = new ResourceCurlGenerateResult();
-        String curl = tryBuildCurlByRule(content.getSourceContent());
-        if (StringUtils.isNotBlank(curl)) {
-            result.setCurl(curl);
-            result.setSource(CURL_GENERATE_SOURCE_RULE);
-            result.setMessage("规则生成成功");
-            return result;
-        }
-
-        curl = buildCurlByLargeModel(content.getSourceContent());
-        validateSafeCurlCommand(curl);
-        result.setCurl(curl);
-        result.setSource(CURL_GENERATE_SOURCE_LLM);
-        result.setMessage("大模型生成成功");
-        return result;
+        return resourceCurlService.generateResourceCurl(request);
     }
 
     /**
@@ -293,491 +236,10 @@ public class ToolManService {
      * @date 2026-05-08 00:00:00
      */
     public ResourceCurlRunResult runResourceCurl(ResourceCurlRunRequest request) {
-        Long resourceId = request == null ? null : request.getResourceId();
-        String curl = normalizeCurlLineContinuation(request == null ? null : request.getCurl());
-        ResourceCurlContent content = loadResourceCurlContent(resourceId);
-        validateSafeCurlCommand(curl);
-
-        ParsedCurl parsed = CurlParser.parse(curl);
-        validateCurlTargetHost(parsed.getFullUrl(), content);
-
-        long start = System.currentTimeMillis();
-        ResourceCurlRunResult result = new ResourceCurlRunResult();
-        try (Response response = RESOURCE_CURL_HTTP_CLIENT.newCall(buildHttpRequest(parsed)).execute()) {
-            result.setSuccess(response.isSuccessful());
-            result.setStatusCode(response.code());
-            result.setHeaders(flattenHeaders(response));
-            ResponseBody responseBody = response.body();
-            result.setBody(responseBody == null ? "" : responseBody.string());
-        } catch (IOException e) {
-            result.setSuccess(false);
-            result.setErrorMessage(e.getMessage());
-        } finally {
-            result.setDurationMs(System.currentTimeMillis() - start);
-        }
-        return result;
+        return resourceCurlService.runResourceCurl(request);
     }
 
-    /**
-     * 加载工具类资源的 sourceContent 与 targetContent。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private ResourceCurlContent loadResourceCurlContent(Long resourceId) {
-        if (resourceId == null) {
-            throw new IllegalArgumentException(I18nUtil.get("resource.resourceid.notnull"));
-        }
-        SsResource resource = ssResourceService.findById(resourceId);
-        if (resource == null) {
-            throw new IllegalArgumentException(I18nUtil.get("resource.notfound"));
-        }
-        String resourceBizType = StringUtils.trimToEmpty(resource.getResourceBizType());
-        if (!TOOL_CURL_BIZ_TYPES.contains(resourceBizType)) {
-            throw new IllegalArgumentException("仅工具类资源支持生成和运行curl脚本");
-        }
-        ResourceCurlContent content = new ResourceCurlContent();
-        content.setResource(resource);
-
-        /**
-         *
-        if (ResourceBizType.TOOL.getCode().equals(resourceBizType)) {
-            SsResExtTool ext = ssResExtToolService.findById(resourceId);
-            content.setSourceContent(ext == null ? null : ext.getSourceContent());
-            content.setTargetContent(ext == null ? null : ext.getTargetContent());
-            return content;
-        }
-         */
-        if (ResourceBizType.TOOLKIT.getCode().equals(resourceBizType)) {
-            SsResExtToolKit ext = ssResExtToolKitService.findById(resourceId);
-            content.setSourceContent(ext == null ? null : ext.getSourceContent());
-            content.setTargetContent(ext == null ? null : ext.getTargetContent());
-            return content;
-        }
-        if (ResourceBizType.MCP.getCode().equals(resourceBizType)) {
-            SsResExtMcp ext = ssResExtMcpService.findById(resourceId);
-            content.setSourceContent(ext == null ? null : ext.getSourceContent());
-            content.setTargetContent(ext == null ? null : ext.getTargetContent());
-            return content;
-        }
-        SsResExtAgent ext = ssResExtAgentService.findById(resourceId);
-        content.setSourceContent(ext == null ? null : ext.getSourceContent());
-        content.setTargetContent(ext == null ? null : ext.getTargetContent());
-        return content;
-    }
-
-    /**
-     * 尝试通过规则从 sourceContent 生成 curl。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String tryBuildCurlByRule(String sourceContent) {
-        String trimmed = StringUtils.trimToEmpty(sourceContent);
-        if (StringUtils.startsWithIgnoreCase(trimmed, "curl ")) {
-            String normalizedCurl = normalizeCurlLineContinuation(trimmed);
-            validateSafeCurlCommand(normalizedCurl);
-            return normalizedCurl;
-        }
-        try {
-            JSONObject openApi = resolveOpenApiNode(JSON.parseObject(trimmed, Feature.OrderedField));
-            if (openApi == null || openApi.isEmpty()) {
-                return null;
-            }
-            String baseUrl = resolveOpenApiBaseUrl(openApi);
-            JSONObject paths = openApi.getJSONObject("paths");
-            if (paths == null || paths.isEmpty()) {
-                return null;
-            }
-            String path = paths.keySet().iterator().next();
-            JSONObject pathItem = paths.getJSONObject(path);
-            String method = resolveFirstHttpMethod(pathItem);
-            if (StringUtils.isBlank(baseUrl) || StringUtils.isBlank(path) || StringUtils.isBlank(method)) {
-                return null;
-            }
-            JSONObject operation = pathItem.getJSONObject(method);
-            String requestBody = buildCurlRequestBody(operation);
-            return buildCurlScript(method, joinUrl(baseUrl, path), requestBody);
-        } catch (Exception e) {
-            LOGGER.info("规则生成curl失败，将尝试大模型兜底, reason={}", e.getMessage());
-            return null;
-        }
-    }
-
-    /**
-     * 从资源 JSON 中解析 OpenAPI 节点。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private JSONObject resolveOpenApiNode(JSONObject root) {
-        if (root == null) {
-            return null;
-        }
-        if (root.containsKey("paths")) {
-            return root;
-        }
-        JSONObject openApi = root.getJSONObject("openAPI");
-        if (openApi != null) {
-            return openApi;
-        }
-        JSONArray tools = root.getJSONArray("tools");
-        if (tools != null && !tools.isEmpty()) {
-            JSONObject firstTool = tools.getJSONObject(0);
-            return firstTool == null ? null : firstTool.getJSONObject("openAPI");
-        }
-        return null;
-    }
-
-    /**
-     * 解析 OpenAPI 基础地址。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String resolveOpenApiBaseUrl(JSONObject openApi) {
-        JSONArray servers = openApi.getJSONArray("servers");
-        if (servers != null && !servers.isEmpty()) {
-            JSONObject server = servers.getJSONObject(0);
-            String url = server == null ? null : server.getString("url");
-            if (StringUtils.isNotBlank(url)) {
-                return url;
-            }
-        }
-        return StringUtils.defaultIfBlank(openApi.getString("domainURL"), openApi.getString("domainUrl"));
-    }
-
-    /**
-     * 解析 paths 节点中的首个 HTTP 方法。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String resolveFirstHttpMethod(JSONObject pathItem) {
-        if (pathItem == null) {
-            return null;
-        }
-        for (String method : List.of("get", "post", "put", "patch", "delete")) {
-            if (pathItem.containsKey(method)) {
-                return method.toUpperCase(Locale.ROOT);
-            }
-        }
-        return null;
-    }
-
-    /**
-     * 根据 OpenAPI operation 生成请求体示例。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String buildCurlRequestBody(JSONObject operation) {
-        if (operation == null) {
-            return null;
-        }
-        JSONObject requestBody = operation.getJSONObject("requestBody");
-        JSONObject content = requestBody == null ? null : requestBody.getJSONObject("content");
-        if (content == null || content.isEmpty()) {
-            return null;
-        }
-        JSONObject media = content.getJSONObject(content.keySet().iterator().next());
-        if (media == null) {
-            return null;
-        }
-        Object example = media.get("example");
-        if (example != null) {
-            return JSON.toJSONString(example);
-        }
-        JSONObject schema = media.getJSONObject("schema");
-        JSONObject properties = schema == null ? null : schema.getJSONObject("properties");
-        if (properties == null || properties.isEmpty()) {
-            return "{}";
-        }
-        JSONObject body = new JSONObject(true);
-        for (String key : properties.keySet()) {
-            JSONObject field = properties.getJSONObject(key);
-            body.put(key, sampleValueBySchema(field));
-        }
-        return body.toJSONString();
-    }
-
-    /**
-     * 根据 JSON Schema 生成字段示例值。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private Object sampleValueBySchema(JSONObject schema) {
-        if (schema == null) {
-            return "";
-        }
-        Object example = schema.get("example");
-        if (example != null) {
-            return example;
-        }
-        String type = StringUtils.defaultIfBlank(schema.getString("type"), "string");
-        return switch (type) {
-            case "integer", "number" -> 0;
-            case "boolean" -> false;
-            case "array" -> new JSONArray();
-            case "object" -> new JSONObject(true);
-            default -> "";
-        };
-    }
-
-    /**
-     * 组装 shell 风格 curl 文本。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String buildCurlScript(String method, String url, String requestBody) {
-        StringBuilder curl = new StringBuilder();
-        curl.append("curl -X ").append(method).append(" ").append(shellQuote(url));
-        curl.append(" \\\n  -H ").append(shellQuote("Content-Type: application/json"));
-        if (StringUtils.isNotBlank(requestBody) && !"GET".equalsIgnoreCase(method)) {
-            curl.append(" \\\n  -d ").append(shellQuote(requestBody));
-        }
-        return curl.toString();
-    }
-
-    /**
-     * 使用大模型从 sourceContent 兜底生成 curl。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String buildCurlByLargeModel(String sourceContent) {
-        String prompt = """
-            你是一个严格的接口调试脚本生成器。请根据下面的 sourceContent 生成一条可以直接用于接口调试的 curl 命令。
-            要求：
-            1. 只输出一条 curl 命令，不要解释，不要思考过程，不要 Markdown 代码块，不要 <think> 标签。
-            2. 不要输出分号、管道、重定向、&&、||、反引号、$() 等 shell 控制符。
-            3. 优先使用 sourceContent 中的真实 domainURL、servers.url、paths、method、headers、requestBody 示例。
-            4. 如果缺少请求体示例，请按 schema 生成最小 JSON 示例。
-            5. 最终回答必须以 curl 开头，并且只能包含这一条命令。
-
-            sourceContent:
-            %s
-            """.formatted(sourceContent);
-        return normalizeCurlFromLargeModel(aiService.generateText(prompt, null));
-    }
-
-    /**
-     * 规范化大模型输出，提取 curl 命令文本。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String normalizeCurlFromLargeModel(String modelOutput) {
-        String curl = StringUtils.trimToEmpty(modelOutput);
-        curl = curl.replaceAll("(?is)<think>.*?</think>", "");
-        curl = curl.replace("```bash", "").replace("```shell", "").replace("```", "").trim();
-        String[] lines = curl.split("\\R");
-        int commandStartLine = -1;
-        for (int i = 0; i < lines.length; i++) {
-            if (StringUtils.startsWithIgnoreCase(StringUtils.trimToEmpty(lines[i]), "curl ")) {
-                commandStartLine = i;
-            }
-        }
-        if (commandStartLine >= 0) {
-            StringBuilder command = new StringBuilder(StringUtils.trimToEmpty(lines[commandStartLine]));
-            for (int i = commandStartLine + 1; i < lines.length; i++) {
-                String nextLine = StringUtils.trimToEmpty(lines[i]);
-                if (StringUtils.isBlank(nextLine)) {
-                    continue;
-                }
-                if (StringUtils.endsWith(command.toString(), "\\")
-                        || StringUtils.startsWith(nextLine, "-")
-                        || StringUtils.startsWith(nextLine, "--")) {
-                    command.append('\n').append(nextLine);
-                    continue;
-                }
-                break;
-            }
-            curl = command.toString();
-        }
-        return normalizeCurlLineContinuation(curl);
-    }
-
-    /**
-     * 去掉 shell 换行续写符，便于后端解析 curl 参数。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String normalizeCurlLineContinuation(String curl) {
-        return StringUtils.trimToEmpty(curl)
-                .replace("\\\r\n", " ")
-                .replace("\\\n", " ")
-                .replace("\\\r", " ");
-    }
-
-    /**
-     * 校验 curl 文本，避免 shell 控制符进入运行链路。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private void validateSafeCurlCommand(String curl) {
-        String trimmed = StringUtils.trimToEmpty(curl);
-        if (StringUtils.isBlank(trimmed)) {
-            throw new IllegalArgumentException("curl脚本不能为空");
-        }
-        if (!StringUtils.startsWithIgnoreCase(trimmed, "curl ")) {
-            throw new IllegalArgumentException("仅支持curl命令");
-        }
-        validateNoUnquotedShellControl(trimmed);
-    }
-
-    /**
-     * 仅拦截未被引号包裹的 shell 控制符，避免 JSON 请求体中的普通字符被误判。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private void validateNoUnquotedShellControl(String curl) {
-        boolean inSingleQuote = false;
-        boolean inDoubleQuote = false;
-        boolean escape = false;
-        for (int i = 0; i < curl.length(); i++) {
-            char current = curl.charAt(i);
-            if (escape) {
-                escape = false;
-                continue;
-            }
-            if (current == '\\' && !inSingleQuote) {
-                escape = true;
-                continue;
-            }
-            if (current == '\'' && !inDoubleQuote) {
-                inSingleQuote = !inSingleQuote;
-                continue;
-            }
-            if (current == '"' && !inSingleQuote) {
-                inDoubleQuote = !inDoubleQuote;
-                continue;
-            }
-            if (inSingleQuote || inDoubleQuote) {
-                continue;
-            }
-            char next = i + 1 < curl.length() ? curl.charAt(i + 1) : '\0';
-            if (current == '`' || (current == '$' && next == '(')) {
-                throw new IllegalArgumentException("curl脚本包含不允许的shell控制符");
-            }
-        }
-    }
-
-    /**
-     * 校验 curl 目标域名必须在当前资源定义范围内。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private void validateCurlTargetHost(String fullUrl, ResourceCurlContent content) {
-        Set<String> allowedHosts = extractHosts(content.getSourceContent());
-        allowedHosts.addAll(extractHosts(content.getTargetContent()));
-        if (allowedHosts.isEmpty()) {
-            return;
-        }
-        String host = parseHost(fullUrl);
-        if (StringUtils.isBlank(host) || !allowedHosts.contains(host)) {
-            throw new IllegalArgumentException("curl目标地址不在当前资源定义范围内");
-        }
-    }
-
-    /**
-     * 从文本中提取 URL host。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private Set<String> extractHosts(String content) {
-        Set<String> hosts = new HashSet<>();
-        if (StringUtils.isBlank(content)) {
-            return hosts;
-        }
-        Matcher matcher = URL_PATTERN.matcher(content);
-        while (matcher.find()) {
-            String host = parseHost(matcher.group());
-            if (StringUtils.isNotBlank(host)) {
-                hosts.add(host);
-            }
-        }
-        return hosts;
-    }
-
-    /**
-     * 解析 URL host。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String parseHost(String url) {
-        try {
-            URI uri = URI.create(url);
-            return StringUtils.lowerCase(uri.getHost());
-        } catch (Exception e) {
-            return null;
-        }
-    }
-
-    /**
-     * 将解析后的 curl 转成 OkHttp 请求。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private Request buildHttpRequest(ParsedCurl parsed) {
-        Request.Builder builder = new Request.Builder().url(parsed.getFullUrl());
-        parsed.getHeaders().forEach(builder::addHeader);
-        RequestBody requestBody = null;
-        if (StringUtils.isNotBlank(parsed.getBody())) {
-            String contentType = parsed.getHeaders().getOrDefault("Content-Type",
-                    parsed.getHeaders().getOrDefault("content-type", "application/json"));
-            requestBody = RequestBody.create(parsed.getBody(), MediaType.parse(contentType));
-        }
-        String method = StringUtils.upperCase(parsed.getMethod());
-        if (requestBody == null && Set.of("POST", "PUT", "PATCH").contains(method)) {
-            requestBody = RequestBody.create("", MediaType.parse("application/octet-stream"));
-        }
-        builder.method(method, requestBody);
-        return builder.build();
-    }
-
-    /**
-     * 将响应头压平成前端容易展示的 Map。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private Map<String, String> flattenHeaders(Response response) {
-        Map<String, String> headers = new LinkedHashMap<>();
-        response.headers().names().forEach(name -> headers.put(name, response.header(name)));
-        return headers;
-    }
-
-    /**
-     * 拼接 baseUrl 与 path。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String joinUrl(String baseUrl, String path) {
-        String safeBase = StringUtils.removeEnd(StringUtils.trimToEmpty(baseUrl), "/");
-        String safePath = StringUtils.prependIfMissing(StringUtils.trimToEmpty(path), "/");
-        return safeBase + safePath;
-    }
-
-    /**
-     * 对 shell 参数做单引号包裹。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private String shellQuote(String value) {
-        return "'" + StringUtils.defaultString(value).replace("'", "'\"'\"'") + "'";
-    }
+    // ==================== 工具保存与 JSON 导入 ====================
 
     /**
      * 阶段二：用户补全描述后保存入库
@@ -1027,6 +489,8 @@ public class ToolManService {
      * 对象超市：上传对象压缩包，批量解析 objects 目录下的一级对象目录并导入。
      */
     @Transactional(rollbackFor = Exception.class)
+    // ==================== 对象/视图 zip 导入入口 ====================
+
     public ObjectZipImportResult importObjectZipFromMultipart(MultipartFile file, Long catalogId, String ownerType) {
         // 1. 校验请求参数，并把前端上传的 zip 原样落盘。
         validateObjectZipRequest(file, ownerType);
@@ -1200,6 +664,8 @@ public class ToolManService {
         result.setZipFileName(storedZipPath.getFileName().toString());
         return result;
     }
+
+    // ==================== 资源生命周期 ====================
 
     /**
      * 删除资源。
@@ -1437,6 +903,8 @@ public class ToolManService {
         updateExtTargetContentAndSync(resource);
     }
 
+    // ==================== 权限与业务校验 ====================
+
     /**
      * 商业版本（dataset.system=WHALE_AGENT）下，知识/工具的编辑/注销需走智能体门户，
      * 在本系统拦截并提示用户。对象、视图、数字员工和非商业版本不受限。
@@ -1480,7 +948,7 @@ public class ToolManService {
         throw new IllegalArgumentException(I18nUtil.get("user.permission.nopermission"));
     }
 
-    // ==================== 导入：参数校验 ====================
+    // ==================== 导入参数与 targetContent 同步 ====================
 
     private String validateAndReadFile(MultipartFile file) {
         if (file == null || file.isEmpty()) {
@@ -1830,16 +1298,7 @@ public class ToolManService {
         }
     }
 
-    private void deleteResourceRelations(Long resourceId) {
-        // 统一删除两类关系：
-        // 1) 当前资源作为主资源(resource_id)的关系
-        // 2) 当前资源作为被关联资源(rel_resource_id)的关系
-        LambdaQueryWrapper<SsResourceRelDetail> wrapper = new LambdaQueryWrapper<>();
-        wrapper.eq(SsResourceRelDetail::getResourceId, resourceId)
-                .or()
-                .eq(SsResourceRelDetail::getRelResourceId, resourceId);
-        ssResourceRelDetailService.remove(wrapper);
-    }
+
 
     private void saveOrUpdateExtObject(SsResExtObject extObject, Long resourceId) {
         SsResExtObject oldExt = ssResExtObjectService.findById(resourceId);
@@ -2019,6 +1478,8 @@ public class ToolManService {
         ext.setTargetContent(targetContent);
         return ext;
     }
+
+    // ==================== 对象/视图 zip 导入内部流程 ====================
 
     private void validateObjectZipRequest(MultipartFile file, String ownerType) {
         if (file == null || file.isEmpty()) {
@@ -2866,24 +2327,9 @@ public class ToolManService {
     /**
      * add by qin.guoquan 2026-03-31
      */
-    private static Long parseLongFlexible(Object raw) {
-        if (raw == null) {
-            return null;
-        }
-        if (raw instanceof Number) {
-            return ((Number) raw).longValue();
-        }
-        String s = String.valueOf(raw).trim();
-        if (s.isEmpty()) {
-            return null;
-        }
-        try {
-            return Long.parseLong(s);
-        }
-        catch (NumberFormatException e) {
-            return null;
-        }
-    }
+
+
+    // ==================== 导入权限、结果与差异对比 ====================
 
     /**
      * 对象、视图导入成功后，自动补齐创建人的管理授权，避免导入人无法在授权侧继续维护资源。
@@ -3153,6 +2599,8 @@ public class ToolManService {
         return JSON.toJSONString(raw, true);
     }
 
+    // ==================== 内部结果对象 ====================
+
     private static final class ObjectImportSaveResult {
         private Long resourceId;
         private Long catalogId;
@@ -3273,42 +2721,6 @@ public class ToolManService {
 
         List<String> getMissingObjectCodes() {
             return missingObjectCodes;
-        }
-    }
-
-    /**
-     * 资源 curl 生成与运行所需的扩展内容载体。
-     *
-     * @author qin.guoquan
-     * @date 2026-05-08 00:00:00
-     */
-    private static final class ResourceCurlContent {
-        private SsResource resource;
-        private String sourceContent;
-        private String targetContent;
-
-        SsResource getResource() {
-            return resource;
-        }
-
-        void setResource(SsResource resource) {
-            this.resource = resource;
-        }
-
-        String getSourceContent() {
-            return sourceContent;
-        }
-
-        void setSourceContent(String sourceContent) {
-            this.sourceContent = sourceContent;
-        }
-
-        String getTargetContent() {
-            return targetContent;
-        }
-
-        void setTargetContent(String targetContent) {
-            this.targetContent = targetContent;
         }
     }
 }
