+++ byclaw-be/src/test/java/com/iwhalecloud/byai/state/application/service/session/ByClawSkillQueryApplicationServiceTest.java
@@ -3,6 +3,8 @@ package com.iwhalecloud.byai.state.application.service.session;
 import com.iwhalecloud.byai.common.i18n.I18nUtil;
 import com.iwhalecloud.byai.common.login.auth.CurrentUserHolder;
 import com.iwhalecloud.byai.common.storage.UserFS;
+import com.iwhalecloud.byai.manager.domain.resource.service.SsResourceService;
+import com.iwhalecloud.byai.manager.entity.resource.SsResource;
 import com.iwhalecloud.byai.state.domain.session.dto.ByClawSkillDto;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -31,11 +33,14 @@ class ByClawSkillQueryApplicationServiceTest {
 
     private static final String USER_CODE = "adminvip";
     private static final Long RESOURCE_ID = 10000417L;
-    private static final String SKILL_ROOT_PREFIX = "/.openclaw/workspace-baiying-agent-10000417/skills/";
+    private static final String AGENT_SKILL_ROOT_PREFIX = "/.openclaw/workspace-baiying-agent-10000417/skills/";
     private static final String WORKSPACE_SKILL_ROOT_PREFIX = "/.openclaw/workspace/skills/";
 
     @Mock
     private UserFS userFS;
+
+    @Mock
+    private SsResourceService ssResourceService;
 
     private ByClawSkillQueryApplicationService byClawSkillQueryApplicationService;
 
@@ -49,6 +54,7 @@ class ByClawSkillQueryApplicationServiceTest {
 
         byClawSkillQueryApplicationService = new ByClawSkillQueryApplicationService();
         ReflectionTestUtils.setField(byClawSkillQueryApplicationService, "userFS", userFS);
+        ReflectionTestUtils.setField(byClawSkillQueryApplicationService, "ssResourceService", ssResourceService);
     }
 
     @AfterEach
@@ -58,41 +64,34 @@ class ByClawSkillQueryApplicationServiceTest {
 
     @Test
     void shouldAggregateSkillDirectoriesAndSortBySkillName() {
-        when(userFS.list(eq(SKILL_ROOT_PREFIX), isNull())).thenReturn(Arrays.asList(
-            SKILL_ROOT_PREFIX + "zeta/SKILL.md",
-            SKILL_ROOT_PREFIX + "zeta/resources/a.txt",
-            SKILL_ROOT_PREFIX + "zeta/requirements.txt",
-            SKILL_ROOT_PREFIX + "alpha/README.md",
-            SKILL_ROOT_PREFIX + "alpha/SKILL.md",
-            SKILL_ROOT_PREFIX + "README.md"));
-        when(userFS.list(eq(WORKSPACE_SKILL_ROOT_PREFIX), isNull())).thenReturn(Arrays.asList(
-            WORKSPACE_SKILL_ROOT_PREFIX + "beta/SKILL.md",
-            WORKSPACE_SKILL_ROOT_PREFIX + "zeta/SKILL.md",
-            WORKSPACE_SKILL_ROOT_PREFIX + "nested/path/SKILL.md"));
+        when(ssResourceService.findById(RESOURCE_ID)).thenReturn(resource("employee_10000417"));
+        when(userFS.list(eq(AGENT_SKILL_ROOT_PREFIX), isNull())).thenReturn(Arrays.asList(
+            AGENT_SKILL_ROOT_PREFIX + "zeta/SKILL.md",
+            AGENT_SKILL_ROOT_PREFIX + "zeta/resources/a.txt",
+            AGENT_SKILL_ROOT_PREFIX + "zeta/requirements.txt",
+            AGENT_SKILL_ROOT_PREFIX + "alpha/README.md",
+            AGENT_SKILL_ROOT_PREFIX + "alpha/SKILL.md",
+            AGENT_SKILL_ROOT_PREFIX + "README.md"));
 
         List<ByClawSkillDto> result = byClawSkillQueryApplicationService.qrySkillListByUserCode(USER_CODE, RESOURCE_ID, null);
 
-        assertEquals(3, result.size());
+        assertEquals(2, result.size());
         assertEquals("alpha", result.get(0).getSkillName());
-        assertEquals(SKILL_ROOT_PREFIX + "alpha", result.get(0).getSkillPath());
-        assertEquals(SKILL_ROOT_PREFIX + "alpha/SKILL.md", result.get(0).getSkillDocObjectKey());
-        assertEquals("beta", result.get(1).getSkillName());
-        assertEquals(WORKSPACE_SKILL_ROOT_PREFIX + "beta", result.get(1).getSkillPath());
-        assertEquals(WORKSPACE_SKILL_ROOT_PREFIX + "beta/SKILL.md", result.get(1).getSkillDocObjectKey());
-        assertEquals("zeta", result.get(2).getSkillName());
-        assertEquals(SKILL_ROOT_PREFIX + "zeta", result.get(2).getSkillPath());
-        assertEquals(SKILL_ROOT_PREFIX + "zeta/SKILL.md", result.get(2).getSkillDocObjectKey());
-        verify(userFS).list(eq(SKILL_ROOT_PREFIX), isNull());
-        verify(userFS).list(eq(WORKSPACE_SKILL_ROOT_PREFIX), isNull());
+        assertEquals(AGENT_SKILL_ROOT_PREFIX + "alpha", result.get(0).getSkillPath());
+        assertEquals(AGENT_SKILL_ROOT_PREFIX + "alpha/SKILL.md", result.get(0).getSkillDocObjectKey());
+        assertEquals("zeta", result.get(1).getSkillName());
+        assertEquals(AGENT_SKILL_ROOT_PREFIX + "zeta", result.get(1).getSkillPath());
+        assertEquals(AGENT_SKILL_ROOT_PREFIX + "zeta/SKILL.md", result.get(1).getSkillDocObjectKey());
+        verify(userFS).list(eq(AGENT_SKILL_ROOT_PREFIX), isNull());
     }
 
     @Test
     void shouldFilterBySkillNameAndIgnoreDirectoryWithoutSkillDoc() {
-        when(userFS.list(eq(SKILL_ROOT_PREFIX), isNull())).thenReturn(Arrays.asList(
-            SKILL_ROOT_PREFIX + "baiying/resources/prompt.txt",
-            SKILL_ROOT_PREFIX + "baiying/requirements.txt",
-            SKILL_ROOT_PREFIX + "github-weekly-commit-collector/SKILL.md"));
-        when(userFS.list(eq(WORKSPACE_SKILL_ROOT_PREFIX), isNull())).thenReturn(Collections.emptyList());
+        when(ssResourceService.findById(RESOURCE_ID)).thenReturn(resource("employee_10000417"));
+        when(userFS.list(eq(AGENT_SKILL_ROOT_PREFIX), isNull())).thenReturn(Arrays.asList(
+            AGENT_SKILL_ROOT_PREFIX + "baiying/resources/prompt.txt",
+            AGENT_SKILL_ROOT_PREFIX + "baiying/requirements.txt",
+            AGENT_SKILL_ROOT_PREFIX + "github-weekly-commit-collector/SKILL.md"));
 
         List<ByClawSkillDto> result = byClawSkillQueryApplicationService.qrySkillListByUserCode(USER_CODE, RESOURCE_ID, "bai");
 
@@ -101,26 +100,26 @@ class ByClawSkillQueryApplicationServiceTest {
 
     @Test
     void shouldReturnOnlyOneLevelSkillDocFiles() {
-        when(userFS.list(eq(SKILL_ROOT_PREFIX), isNull())).thenReturn(Arrays.asList(
-            SKILL_ROOT_PREFIX + "baiying/SKILL.md",
-            SKILL_ROOT_PREFIX + "baiying/resources/SKILL.md",
-            SKILL_ROOT_PREFIX + "baiying/README.md",
-            SKILL_ROOT_PREFIX + "nested/path/SKILL.md"));
-        when(userFS.list(eq(WORKSPACE_SKILL_ROOT_PREFIX), isNull())).thenReturn(Collections.emptyList());
+        when(ssResourceService.findById(RESOURCE_ID)).thenReturn(resource("employee_10000417"));
+        when(userFS.list(eq(AGENT_SKILL_ROOT_PREFIX), isNull())).thenReturn(Arrays.asList(
+            AGENT_SKILL_ROOT_PREFIX + "baiying/SKILL.md",
+            AGENT_SKILL_ROOT_PREFIX + "baiying/resources/SKILL.md",
+            AGENT_SKILL_ROOT_PREFIX + "baiying/README.md",
+            AGENT_SKILL_ROOT_PREFIX + "nested/path/SKILL.md"));
 
         List<ByClawSkillDto> result = byClawSkillQueryApplicationService.qrySkillListByUserCode(USER_CODE, RESOURCE_ID, null);
 
         assertEquals(1, result.size());
         assertEquals("baiying", result.get(0).getSkillName());
-        assertEquals(SKILL_ROOT_PREFIX + "baiying/SKILL.md", result.get(0).getSkillDocObjectKey());
+        assertEquals(AGENT_SKILL_ROOT_PREFIX + "baiying/SKILL.md", result.get(0).getSkillDocObjectKey());
     }
 
     @Test
     void shouldMatchKeywordAgainstSkillDirectoryNameOnly() {
-        when(userFS.list(eq(SKILL_ROOT_PREFIX), isNull())).thenReturn(Arrays.asList(
-            SKILL_ROOT_PREFIX + "baiying-agent/SKILL.md",
-            SKILL_ROOT_PREFIX + "other-skill/SKILL.md"));
-        when(userFS.list(eq(WORKSPACE_SKILL_ROOT_PREFIX), isNull())).thenReturn(Collections.emptyList());
+        when(ssResourceService.findById(RESOURCE_ID)).thenReturn(resource("employee_10000417"));
+        when(userFS.list(eq(AGENT_SKILL_ROOT_PREFIX), isNull())).thenReturn(Arrays.asList(
+            AGENT_SKILL_ROOT_PREFIX + "baiying-agent/SKILL.md",
+            AGENT_SKILL_ROOT_PREFIX + "other-skill/SKILL.md"));
 
         List<ByClawSkillDto> result = byClawSkillQueryApplicationService.qrySkillListByUserCode(USER_CODE, RESOURCE_ID, "agent");
 
@@ -130,10 +129,8 @@ class ByClawSkillQueryApplicationServiceTest {
 
     @Test
     void shouldReturnEmptyListWhenBucketOrPrefixHasNoObjects() {
-        when(userFS.list(eq(SKILL_ROOT_PREFIX), isNull()))
-            .thenReturn(Collections.emptyList());
-        when(userFS.list(eq(WORKSPACE_SKILL_ROOT_PREFIX), isNull()))
-            .thenReturn(Collections.emptyList());
+        when(ssResourceService.findById(RESOURCE_ID)).thenReturn(resource("employee_10000417"));
+        when(userFS.list(eq(AGENT_SKILL_ROOT_PREFIX), isNull())).thenReturn(Collections.emptyList());
 
         List<ByClawSkillDto> result = byClawSkillQueryApplicationService.qrySkillListByUserCode(USER_CODE, RESOURCE_ID, null);
 
@@ -142,10 +139,8 @@ class ByClawSkillQueryApplicationServiceTest {
 
     @Test
     void shouldReturnEmptyListWhenObjectKeysIsNull() {
-        when(userFS.list(eq(SKILL_ROOT_PREFIX), isNull()))
-            .thenReturn(null);
-        when(userFS.list(eq(WORKSPACE_SKILL_ROOT_PREFIX), isNull()))
-            .thenReturn(null);
+        when(ssResourceService.findById(RESOURCE_ID)).thenReturn(resource("employee_10000417"));
+        when(userFS.list(eq(AGENT_SKILL_ROOT_PREFIX), isNull())).thenReturn(null);
 
         List<ByClawSkillDto> result = byClawSkillQueryApplicationService.qrySkillListByUserCode(USER_CODE, RESOURCE_ID, null);
 
@@ -161,16 +156,8 @@ class ByClawSkillQueryApplicationServiceTest {
     }
 
     @Test
-    void shouldRejectNullResourceId() {
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-            () -> byClawSkillQueryApplicationService.qrySkillListByUserCode(USER_CODE, null, null));
-
-        assertEquals("资源ID不能为空", exception.getMessage());
-    }
-
-    @Test
-    void shouldIncludeGlobalWorkspaceSkillsWhenUserWorkspaceIsEmpty() {
-        when(userFS.list(eq(SKILL_ROOT_PREFIX), isNull())).thenReturn(Collections.emptyList());
+    void shouldQuerySuperAssistantWorkspaceWhenResourceIdIsNull() {
+        when(ssResourceService.findById(RESOURCE_ID)).thenReturn(resource("adminvip_main"));
         when(userFS.list(eq(WORKSPACE_SKILL_ROOT_PREFIX), isNull())).thenReturn(Arrays.asList(
             WORKSPACE_SKILL_ROOT_PREFIX + "assistant-core/SKILL.md",
             WORKSPACE_SKILL_ROOT_PREFIX + "assistant-core/resources/logo.png"));
@@ -181,5 +168,11 @@ class ByClawSkillQueryApplicationServiceTest {
         assertEquals("assistant-core", result.get(0).getSkillName());
         assertEquals(WORKSPACE_SKILL_ROOT_PREFIX + "assistant-core", result.get(0).getSkillPath());
         assertEquals(WORKSPACE_SKILL_ROOT_PREFIX + "assistant-core/SKILL.md", result.get(0).getSkillDocObjectKey());
+    }
+
+    private SsResource resource(String resourceCode) {
+        SsResource resource = new SsResource();
+        resource.setResourceCode(resourceCode);
+        return resource;
     }
 }
